requirements.txt
click
click_shell
grpcio
googleapis-common-protos
grpcio-status
grpcio_tools
kafka-python
nest_asyncio
rich
requests
flask
flask-restful
sh
druncschema @ git+https://github.com/DUNE-DAQ/druncschema.git
